Move domain models into the domain package.
