+++ slackhealthbot/domain/models/sleep.py
@@ -1,0 +1,10 @@
+import datetime
+
+from pydantic import BaseModel, NonNegativeInt
+
+
+class SleepData(BaseModel):
+    start_time: datetime.datetime
+    end_time: datetime.datetime
+    sleep_minutes: NonNegativeInt
+    wake_minutes: NonNegativeInt
