+++ slackhealthbot/domain/usecases/fitbit/usecase_process_new_activity.py
@@ -2,13 +2,13 @@ import datetime
 
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from slackhealthbot.core.models import ActivityData, ActivityHistory
-from slackhealthbot.domain.modelmappers.coretorepository.activity import (
+from slackhealthbot.domain.modelmappers.domaintorepository.activity import (
     core_activity_to_repository_activity,
 )
-from slackhealthbot.domain.modelmappers.repositorytocore.activity import (
+from slackhealthbot.domain.modelmappers.repositorytodomain.activity import (
     repository_activity_to_core_activity,
 )
+from slackhealthbot.domain.models.activity import ActivityData, ActivityHistory
 from slackhealthbot.domain.usecases.fitbit import usecase_get_last_activity
 from slackhealthbot.domain.usecases.slack import usecase_post_activity
 from slackhealthbot.repositories import fitbitrepository
