+++ slackhealthbot/domain/usecases/fitbit/usecase_process_new_sleep.py
@@ -2,10 +2,10 @@ import datetime
 
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from slackhealthbot.core.models import SleepData
-from slackhealthbot.domain.modelmappers.coretorepository.sleep import (
+from slackhealthbot.domain.modelmappers.domaintorepository.sleep import (
     core_sleep_to_repository_sleep,
 )
+from slackhealthbot.domain.models.sleep import SleepData
 from slackhealthbot.domain.usecases.fitbit import usecase_get_last_sleep
 from slackhealthbot.domain.usecases.slack import usecase_post_sleep
 from slackhealthbot.repositories import fitbitrepository
