+++ slackhealthbot/domain/usecases/slack/usecase_post_sleep.py
@@ -2,7 +2,7 @@ import datetime
 
 import httpx
 
-from slackhealthbot.core.models import SleepData
+from slackhealthbot.domain.models.sleep import SleepData
 from slackhealthbot.settings import settings
