+++ tests/services/fitbit/test_parser.py
@@ -4,10 +4,10 @@ from pathlib import Path
 
 import pytest
 
-from slackhealthbot.core.models import SleepData
-from slackhealthbot.domain.modelmappers.remoteservicetocore.sleep import (
+from slackhealthbot.domain.modelmappers.remoteservicetodomain.sleep import (
     remote_service_sleep_to_core_sleep,
 )
+from slackhealthbot.domain.models.sleep import SleepData
 from slackhealthbot.remoteservices.fitbit.sleepapi import FitbitSleep
